third_party/blink/web_tests/html/document_metadata/base-expected.txt

PASS endsWith(document.querySelector('base').href, 'foo/bar/') is true
PASS endsWith(document.querySelector('base').href, 'foo/bar/foo/bar/') is false
PASS endsWith(document.querySelector('base').href, '/null') is true
PASS endsWith(documentWithoutAView.querySelector('base').href, 'foo/bar/') is true
PASS documentWithoutAView.querySelector('base').href is "http://webkit.org/"
PASS endsWith(iframe.contentDocument.querySelector('base').href, 'foo/bar/') is true
PASS newWindow.document.querySelector('base').href is ""
PASS successfullyParsed is true

TEST COMPLETE
